+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -9,7 +9,9 @@ import { SortOrder } from '@/common/query-filters/sort-order';
 import { TDocumentsWithoutPageType, TDocumentWithoutPageType } from '@/common/types';
 import { aliasIndividualAsEndUser } from '@/common/utils/alias-individual-as-end-user/alias-individual-as-end-user';
 import { logDocumentWithoutId } from '@/common/utils/log-document-without-id/log-document-without-id';
+import { TOcrImages, UnifiedApiClient } from '@/common/utils/unified-api-client/unified-api-client';
 import { CustomerService } from '@/customer/customer.service';
+import { FEATURE_LIST } from '@/customer/types';
 import { EndUserRepository } from '@/end-user/end-user.repository';
 import { EndUserService } from '@/end-user/end-user.service';
 import { env } from '@/env';
@@ -24,8 +26,11 @@ import { FileService } from '@/providers/file/file.service';
 import { RiskRuleService, TFindAllRulesOptions } from '@/rule-engine/risk-rule.service';
 import { RuleEngineService } from '@/rule-engine/rule-engine.service';
 import { SalesforceService } from '@/salesforce/salesforce.service';
+import { AwsSecretsManager } from '@/secrets-manager/aws-secrets-manager';
+import { InMemorySecretsManager } from '@/secrets-manager/in-memory-secrets-manager';
 import { SecretsManagerFactory } from '@/secrets-manager/secrets-manager.factory';
 import { SentryService } from '@/sentry/sentry.service';
+import { StorageService } from '@/storage/storage.service';
 import type {
   InputJsonValue,
   IObjectWithId,
@@ -47,11 +52,15 @@ import {
 } from '@/workflow/workflow-runtime-list-item.model';
 import {
   AnyRecord,
+  buildCollectionFlowState,
+  CollectionFlowStatusesEnum,
   DefaultContextSchema,
   getDocumentId,
+  getOrderedSteps,
   isErrorWithMessage,
   isObject,
   ProcessStatus,
+  setCollectionFlowStatus,
 } from '@ballerine/common';
 import {
   ARRAY_MERGE_OPTION,
@@ -93,6 +102,7 @@ import { plainToClass } from 'class-transformer';
 import dayjs from 'dayjs';
 import { isEqual, merge } from 'lodash';
 import mime from 'mime';
+import { WORKFLOW_TERMINAL_STATES } from './consts';
 import { WorkflowDefinitionCreateDto } from './dtos/workflow-definition-create';
 import { WorkflowDefinitionFindManyArgs } from './dtos/workflow-definition-find-many-args';
 import { WorkflowDefinitionUpdateInput } from './dtos/workflow-definition-update-input';
@@ -108,15 +118,18 @@ import { addPropertiesSchemaToDocument } from './utils/add-properties-schema-to-
 import { entitiesUpdate } from './utils/entities-update';
 import { WorkflowEventEmitterService } from './workflow-event-emitter.service';
 import { WorkflowRuntimeDataRepository } from './workflow-runtime-data.repository';
-import { StorageService } from '@/storage/storage.service';
-import { TOcrImages, UnifiedApiClient } from '@/common/utils/unified-api-client/unified-api-client';
-import { AwsSecretsManager } from '@/secrets-manager/aws-secrets-manager';
-import { InMemorySecretsManager } from '@/secrets-manager/in-memory-secrets-manager';
-import { FEATURE_LIST } from '@/customer/types';
 
 type TEntityId = string;
 
 export type TEntityType = 'endUser' | 'business';
+
+type CollectionFlowEvent = 'approved' | 'rejected' | 'failed' | 'revision';
+const COLLECTION_FLOW_EVENTS_WHITELIST: readonly CollectionFlowEvent[] = [
+  'approved',
+  'rejected',
+  'failed',
+  'revision',
+] as const;
 
 @Injectable()
 export class WorkflowService {
@@ -1463,31 +1476,6 @@ export class WorkflowService {
           }
         }
 
-        const uiSchema = (uiDefinition as Record<string, any>)?.uiSchema;
-
-        const createFlowConfig = (uiSchema: Record<string, any>) => {
-          return {
-            stepsProgress: (
-              uiSchema?.elements as Array<{
-                type: string;
-                number: number;
-                stateName: string;
-              }>
-            )?.reduce((acc, curr) => {
-              if (curr?.type !== 'page') {
-                return acc;
-              }
-
-              acc[curr?.stateName] = {
-                number: curr?.number,
-                isCompleted: false,
-              };
-
-              return acc;
-            }, {} as { [key: string]: { number: number; isCompleted: boolean } }),
-          };
-        };
-
         workflowRuntimeData = await this.workflowRuntimeDataRepository.create(
           {
             data: {
@@ -1496,7 +1484,6 @@ export class WorkflowService {
               context: {
                 ...contextToInsert,
                 documents: documentsWithPersistedImages,
-                flowConfig: (contextToInsert as any)?.flowConfig ?? createFlowConfig(uiSchema),
                 metadata: {
                   customerId: customer.id,
                   customerNormalizedName: customer.name,
@@ -1569,19 +1556,33 @@ export class WorkflowService {
             transaction,
           );
 
+          const collectionFlow = buildCollectionFlowState({
+            apiUrl: env.APP_API_URL,
+            steps: getOrderedSteps(
+              (uiDefinition?.definition as Prisma.JsonObject)?.definition as Record<string, any>,
+              { terminalStates: [...WORKFLOW_TERMINAL_STATES] },
+            ).map(stepName => ({
+              stateName: stepName,
+            })),
+            additionalInformation: {
+              customerCompany: customer.displayName,
+            },
+          });
+
           workflowRuntimeData = await this.workflowRuntimeDataRepository.updateStateById(
             workflowRuntimeData.id,
             {
               data: {
                 context: {
                   ...workflowRuntimeData.context,
+                  collectionFlow,
                   metadata: {
                     ...(workflowRuntimeData.context.metadata ?? {}),
                     token: workflowToken.token,
                     collectionFlowUrl: env.COLLECTION_FLOW_URL,
                     webUiSDKUrl: env.WEB_UI_SDK_URL,
                   },
-                },
+                } as InputJsonValue,
                 projectId: currentProjectId,
               },
             },
@@ -2175,6 +2176,25 @@ export class WorkflowService {
       const snapshot = service.getSnapshot();
       const currentState = snapshot.value;
       const context = snapshot.machine?.context;
+
+      // Checking if event type is candidate for "revision" state
+      const nextCollectionFlowState = COLLECTION_FLOW_EVENTS_WHITELIST.includes(type)
+        ? type
+        : // Using current state of workflow for approved, rejected, failed
+        COLLECTION_FLOW_EVENTS_WHITELIST.includes(currentState)
+        ? currentState
+        : undefined;
+
+      this.logger.log('Next collection flow state', {
+        nextCollectionFlowState: nextCollectionFlowState || 'N/A',
+      });
+
+      if (nextCollectionFlowState) {
+        if (currentState in CollectionFlowStatusesEnum) {
+          setCollectionFlowStatus(context, currentState);
+        }
+      }
+
       // TODO: Refactor to use snapshot.done instead
       // @ts-ignore
       const isFinal = snapshot.machine?.states[currentState].type === 'final';