+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @ballerine/workflow-node-sdk
 
+## 0.6.61
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.61
+
+## 0.6.60
+
+### Patch Changes
+
+- Change
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.60
+
+## 0.6.59
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.59
+
+## 0.6.58
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.58
+
+## 0.6.57
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.57
+
+## 0.6.56
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.6.56
+
 ## 0.6.55
 
 ### Patch Changes